Kill preprocessor module in block-preprocessor and add some docs
`domain_block_preprocessor::foo` looks a bit nicer than `domain_block_preprocessor::preprocessor::foo`.
